refactor: move package location
