+++ src/main/kotlin/com/oneplace/dgapiserver/domain/account/exception/InvalidFirebaseTokenException.kt
@@ -1,6 +1,6 @@
 package com.oneplace.dgapiserver.domain.account.exception
 
 import com.oneplace.dgapiserver.global.error.BusinessException
-import com.oneplace.dgapiserver.global.error.ErrorCode
+import com.oneplace.dgapiserver.global.error.account.ErrorCode
 
 class InvalidFirebaseTokenException : BusinessException(ErrorCode.INVALID_FIREBASE_TOKEN)
